Add further tests to TransactionBuilderInputState
